Move Contact out of post folder into it's own folder
